Splits extension helpers between Domain and Core projects
