+++ src/i18n/Concrete/UrlLocalizer.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Web;
+using i18n.Helpers;
 
 namespace i18n
 {
